Move ThemeProvider to App component
